feat: move CMP v2 exports to avoid name collisions
BREAKING CHANGE: With CMP v3 support, the old v2-specific exports
are moved to `@s-group/react-usercentrics/v2`. The root
`@s-group/react-usercentrics` only exports generic or v3-specific things.
